Don't show data preview if dataset is large
Refuse to display preview data if the dataset is larger than 1000
records.  Instead show an error message saying "data is too large to
preview, please download".

The ISP project found that after about 3000 responses, the memory
usage of the browser tab was climbing to 1.5Gb.  The actual JSON data
was much smaller (about ~80Mb), but all the decoration added by Ember
in the preview window caused it to explode.

[SVCS-434]
Closes: #145